refactor :  네이밍 개선 및 구조 수정
